backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.12
websockets==13.1
python-dotenv==1.0.1
google-genai==1.14.0
mujoco>=3.1.3